reading-demo/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
openai
azure-search-documents
